Consolidate aleph.ext
